do mvc: add domain directory
