--- src/errors/DataNotFound.js
@@ -1,3 +1,0 @@
-class DataNotFound extends Error {}
-
-module.exports = DataNotFound
