+++ src/errors/DataNotFoundError.js
@@ -1,0 +1,3 @@
+class DataNotFoundError extends Error {}
+
+module.exports = DataNotFoundError
