+++ src/server.js
@@ -10,9 +10,9 @@ app.use(express.json())
 app.use(morgan("dev"))
 
 // REQUIRE ROUTERS
-const usersRouter = require("./routers/users")
-const filmsRouter = require("./routers/films")
-const booksRouter = require("./routers/books")
+const usersRouter = require("./routers/users/users.js")
+const filmsRouter = require("./routers/films/films.js")
+const booksRouter = require("./routers/books/books.js")
 
 // ADD ROUTERS TO APP
 app.use("/users", usersRouter)
@@ -20,9 +20,9 @@ app.use("/films", filmsRouter)
 app.use("/books", booksRouter)
 
 // ADD ERRORS CLASSES
-const MissingDataError = require("./errors/MissingDataError")
-const ExistingDataError = require("./errors/ExistingDataError")
-const DataNotFound = require("./errors/DataNotFound")
+const MissingDataError = require("./errors/MissingDataError.js")
+const ExistingDataError = require("./errors/ExistingDataError.js")
+const DataNotFoundError = require("./errors/DataNotFoundError.js")
 
 app.use((error, req, res, next) => {
   if (error instanceof MissingDataError) {
@@ -37,7 +37,7 @@ app.use((error, req, res, next) => {
     })
   }
 
-  if (error instanceof DataNotFound) {
+  if (error instanceof DataNotFoundError) {
     return res.status(404).json({
       error: error.message,
     })
